refactor: ckBTC sync/load tokens and accounts to wallet
