+++ frontend/src/lib/services/ckbtc-convert.services.ts
@@ -35,10 +35,7 @@ import { approveTransfer } from "../api/icrc-ledger.api";
 import { toastsError } from "../stores/toasts.store";
 import { numberToE8s } from "../utils/token.utils";
 import { getAuthenticatedIdentity } from "./auth.services";
-import {
-  ckBTCTransferTokens,
-  loadCkBTCAccounts,
-} from "./ckbtc-accounts.services";
+import { ckBTCTransferTokens } from "./ckbtc-accounts.services";
 import type { IcrcTransferTokensUserParams } from "./icrc-accounts.services";
 import { loadWalletTransactions } from "./wallet-transactions.services";
 
@@ -299,7 +296,7 @@ const reload = async ({
   // - if provided, the transactions of the account for which the transfer was executed
   // - the balance of the withdrawal account to display an information if some funds - from this transaction or another - are stuck and not been converted yet
   await Promise.all([
-    ...(loadAccounts ? [loadCkBTCAccounts({ universeId })] : []),
+    ...(loadAccounts ? [loadAccounts({ universeId })] : []),
     ...(nonNullish(source)
       ? [
           loadWalletTransactions({
